caches por fin en lado servidor

+++ c_modules/mediaserver/classes/controller/MediaserverController.php
@@ -77,8 +77,8 @@ class MediaserverController {
   */
   function copyAndMoveFile( $minify = false ) {
 
-    $tmp_cache = MEDIASERVER_TMP_CACHE_PATH . $this->modulePath . $this->urlPath;
-    $final_cache = SITE_PATH.'../httpdocs/'.MEDIASERVER_FINAL_CACHE_PATH . $this->modulePath . $this->urlPath;
+    $tmp_cache = MEDIASERVER_TMP_CACHE_PATH .'/'. $this->modulePath . $this->urlPath;
+    $final_cache = SITE_PATH.'../httpdocs/'.MEDIASERVER_FINAL_CACHE_PATH .'/'. $this->modulePath . $this->urlPath;
 
 
     if( !file_exists( $tmp_cache && MEDIASERVER_HOST == '/' ) ) {
@@ -116,26 +116,27 @@ class MediaserverController {
   function compileAndMoveLessFile( $minify = false ) {
 
     $lessControl = new LessController();
-    $tmp_cache = MEDIASERVER_TMP_CACHE_PATH . $this->modulePath . $this->urlPath. '.css';
-    $final_cache = SITE_PATH.'../httpdocs/'.MEDIASERVER_FINAL_CACHE_PATH . $this->modulePath . $this->urlPath .'.css' ;
+    $tmp_cache = MEDIASERVER_TMP_CACHE_PATH .'/'. $this->modulePath . $this->urlPath. '.css';
+    $final_cache = SITE_PATH.'../httpdocs/'.MEDIASERVER_FINAL_CACHE_PATH .'/'. $this->modulePath . $this->urlPath .'.css' ;
 
     // create tmp folder
     $this->createDirPath( $tmp_cache );
 
 
-    $lessControl->compile( $this->urlPath, $tmp_cache, $this->moduleName );
+    if( $lessControl->compile( $this->urlPath, $tmp_cache, $this->moduleName ) ) {
+      // create final folder
+      $this->createDirPath( $final_cache );
 
+      // delete final file if exist
+      if( file_exists( $final_cache ) ){
+        unlink( $final_cache );
+      }
 
-    // create final folder
-    $this->createDirPath( $final_cache );
-
-    // delete final file if exist
-    if( file_exists( $final_cache ) ){
-      unlink( $final_cache );
+      // move from tmp path to final path
+      rename( $tmp_cache , $final_cache );
     }
 
-    // move from tmp path to final path
-    rename( $tmp_cache , $final_cache );
+
   }
